backend/requirements.txt
Flask
Flask-CORS
opencv-python
deepface
spotipy
SQLAlchemy
Flask-SQLAlchemy
Pillow
python-dotenv
tensorflow
PyJWT
psycopg2-binary

# Testing
pytest
pytest-cov